feat(settings): split Settings by scope — the project's, and Frame's own
One gear at the foot of the sidebar opened one modal holding both kinds of
setting, so "Remove Frame from this project" sat a scroll away from "Send
anonymous usage stats" as though they were the same kind of choice. They are
not: one writes into the open project's .frame/ and dies with it, the other
is true of this machine whichever project is open.

Two surfaces now. Project Settings keeps the Workflow section — Spec-Driven
Development, Git sharing, Remove Frame — and opens from the rail's foot,
where the old gear was. Frame Settings holds Privacy & Analytics and About,
and opens from a new button at the right end of the sidebar header, beside
the update dot that already led into About.

The marks had to differ or the split would only move the confusion. The gear
went up: it means application preferences everywhere else, so it belongs to
the app-wide surface. The project's scope wears sliders instead.

settingsModal.js became three modules — one per surface, plus settingsOverlay
for the box they share (backdrop, ×, Escape, re-read on open). The overlay
layer also keeps the two from stacking: the buttons are behind the backdrop
while a dialog is up, but Cmd+, is not.

Every other entry point was re-pointed by scope: Cmd+,, the app menu's
Settings item and the telemetry notice go to Frame Settings; the update dot
and banner go there too, into About. The palette's single "Open Settings"
became "Frame Settings" and "Project Settings".

One quieter thread: specDrivenHint anchors its popover on the button whose
modal holds the switch it is describing. That button is now Project Settings,
so the anchor moved with it — otherwise the hint would have pointed at a
gear that no longer opens what it was talking about.
