+++ src/renderer/frameSettingsModal.js
@@ -1,0 +1,290 @@
+/**
+ * Frame Settings
+ *
+ * Frame's own settings, opened by the gear in the sidebar header (and Cmd+,,
+ * and the app menu's Settings item). Machine-wide: privacy choices and the
+ * About panel with the updater are the same whichever project is open, and
+ * they outlive every one of them.
+ *
+ * The project's own settings are a separate surface (projectSettingsModal),
+ * reached from the sliders button at the foot of the sidebar rail. The gear
+ * is here because a gear means application preferences everywhere else; the
+ * project's scope wears a different mark for the same reason.
+ *
+ * The sidebar's update dot and banner lead here — into About, which is where
+ * the release they are announcing can actually be read.
+ */
+
+const { ipcRenderer, shell } = require('electron');
+const { IPC } = require('../shared/ipcChannels');
+const settingsOverlay = require('./settingsOverlay');
+
+const TELEMETRY_KEY = 'telemetryEnabled';
+const CRASH_DUMPS_KEY = 'crashDumpsEnabled';
+const DISMISSED_VERSION_KEY = 'dismissedUpdateVersion';
+
+let overlay = null;
+let toggleEl = null;
+let crashDumpsToggleEl = null;
+
+// About section elements + state
+let aboutVersionEl = null;
+let aboutStatusEl = null;
+let aboutCheckBtn = null;
+let updateBanner = null;
+let updateLatestEl = null;
+let updateReleasedEl = null;
+let updateLinkEl = null;
+let updateDismissBtn = null;
+let currentUpdateInfo = null;
+
+function init() {
+  toggleEl = document.getElementById('settings-telemetry-toggle');
+  crashDumpsToggleEl = document.getElementById('settings-crash-dumps-toggle');
+
+  // About section
+  aboutVersionEl = document.getElementById('settings-version');
+  aboutStatusEl = document.getElementById('settings-update-status');
+  aboutCheckBtn = document.getElementById('settings-check-updates');
+  updateBanner = document.getElementById('settings-update-available');
+  updateLatestEl = document.getElementById('settings-update-latest');
+  updateReleasedEl = document.getElementById('settings-update-released');
+  updateLinkEl = document.getElementById('settings-update-link');
+  updateDismissBtn = document.getElementById('settings-update-dismiss');
+
+  overlay = settingsOverlay.create('frame-settings-overlay', syncToggleFromSettings);
+  if (!overlay || !toggleEl) {
+    if (!toggleEl) console.error('Frame settings: required elements not found');
+    return;
+  }
+
+  // Load current value
+  syncToggleFromSettings();
+  initAboutSection();
+
+  // Toggle: persist + tell main process to enable/disable Aptabase
+  toggleEl.addEventListener('change', async () => {
+    const enabled = toggleEl.checked;
+    await ipcRenderer.invoke(IPC.SET_USER_SETTING, TELEMETRY_KEY, enabled);
+    await ipcRenderer.invoke(IPC.TELEMETRY_SET_ENABLED, enabled);
+  });
+
+  // Crash dumps: persisted setting only — crashGuard reads it at startup
+  // (the reporter can't be stopped once started, so changes apply on next launch)
+  if (crashDumpsToggleEl) {
+    crashDumpsToggleEl.addEventListener('change', async () => {
+      await ipcRenderer.invoke(IPC.SET_USER_SETTING, CRASH_DUMPS_KEY, crashDumpsToggleEl.checked);
+    });
+  }
+
+  // Open from menu trigger — the app menu's Settings item is application
+  // preferences, so it lands here rather than on the project's panel.
+  ipcRenderer.on(IPC.OPEN_SETTINGS, () => overlay.open());
+
+  // Push updates from periodic recheck refresh the About panel state
+  ipcRenderer.on(IPC.UPDATE_AVAILABLE, (event, info) => {
+    currentUpdateInfo = info;
+    renderUpdateState({ checked: true, found: true, info });
+  });
+}
+
+function initAboutSection() {
+  // Version text from package.json
+  try {
+    const pkgVersion = require('../../package.json').version;
+    if (aboutVersionEl) aboutVersionEl.textContent = `v${pkgVersion}`;
+  } catch (e) { /* ignore */ }
+
+  if (aboutCheckBtn) {
+    aboutCheckBtn.addEventListener('click', () => {
+      runCheck(true);
+    });
+  }
+
+  // Diagnostics: reveal the rotating (redacted) log file — the first thing
+  // to grab when filing a bug report. Location documented in PRIVACY.md.
+  const openLogsBtn = document.getElementById('settings-open-logs');
+  if (openLogsBtn) {
+    openLogsBtn.addEventListener('click', async () => {
+      try {
+        const info = await ipcRenderer.invoke(IPC.GET_LOG_INFO);
+        if (info && info.logPath) shell.showItemInFolder(info.logPath);
+        else if (info && info.logsDir) shell.openPath(info.logsDir);
+      } catch (err) {
+        console.error('Settings: could not open logs folder', err);
+      }
+    });
+  }
+
+  if (updateLinkEl) {
+    updateLinkEl.addEventListener('click', (e) => {
+      e.preventDefault();
+      if (currentUpdateInfo && currentUpdateInfo.releaseUrl) {
+        shell.openExternal(currentUpdateInfo.releaseUrl);
+      }
+    });
+  }
+
+  if (updateDismissBtn) {
+    updateDismissBtn.addEventListener('click', async () => {
+      if (!currentUpdateInfo) return;
+      await ipcRenderer.invoke(
+        IPC.SET_USER_SETTING,
+        DISMISSED_VERSION_KEY,
+        currentUpdateInfo.latestVersion
+      );
+      // Hide the banner immediately; sidebar dot is also gated by this flag.
+      hideUpdateBanner();
+      hideSidebarDot();
+    });
+  }
+
+  // Hydrate from main's cached status (no extra network call)
+  ipcRenderer
+    .invoke(IPC.GET_UPDATE_STATUS)
+    .then((status) => {
+      if (!status) return;
+      if (status.result) {
+        currentUpdateInfo = status.result;
+        renderUpdateState({
+          checked: !!status.lastCheckedAt,
+          found: true,
+          info: status.result,
+          checkedAt: status.lastCheckedAt
+        });
+      } else if (status.lastStatus === 'error') {
+        renderUpdateState({
+          checked: true,
+          failed: true,
+          failReason: status.lastErrorReason,
+          checkedAt: status.lastCheckedAt
+        });
+      } else if (status.lastCheckedAt) {
+        renderUpdateState({
+          checked: true,
+          found: false,
+          checkedAt: status.lastCheckedAt
+        });
+      } else {
+        // Not yet checked since launch — fire one to populate
+        runCheck(false);
+      }
+    })
+    .catch(() => {});
+}
+
+async function runCheck(userInitiated) {
+  if (aboutStatusEl) aboutStatusEl.textContent = 'Checking…';
+  if (aboutCheckBtn) aboutCheckBtn.disabled = true;
+  try {
+    // Discriminated result: 'update-available' | 'up-to-date' | 'error'.
+    // A failed check must never render as "you're up to date".
+    const res = await ipcRenderer.invoke(IPC.CHECK_FOR_UPDATE);
+    if (res && res.status === 'update-available') {
+      currentUpdateInfo = res.info;
+      renderUpdateState({
+        checked: true,
+        found: true,
+        info: res.info,
+        checkedAt: res.checkedAt,
+        userInitiated
+      });
+    } else if (res && res.status === 'error') {
+      renderUpdateState({
+        checked: true,
+        failed: true,
+        failReason: res.reason,
+        checkedAt: res.checkedAt,
+        userInitiated
+      });
+    } else {
+      renderUpdateState({
+        checked: true,
+        found: false,
+        checkedAt: res ? res.checkedAt : null,
+        userInitiated
+      });
+    }
+  } catch (err) {
+    if (aboutStatusEl) aboutStatusEl.textContent = 'Could not check for updates.';
+  } finally {
+    if (aboutCheckBtn) aboutCheckBtn.disabled = false;
+  }
+}
+
+function renderUpdateState({ checked, found, failed, failReason, info, checkedAt, userInitiated }) {
+  if (!aboutStatusEl) return;
+  const stamp = checkedAt ? formatRelative(new Date(checkedAt)) : '';
+  if (found && info) {
+    aboutStatusEl.textContent = stamp ? `Last checked ${stamp}.` : '';
+    showUpdateBanner(info);
+  } else if (failed) {
+    const why = failReason === 'timeout'
+      ? 'timed out'
+      : failReason === 'parse'
+        ? 'unexpected response'
+        : 'network error';
+    aboutStatusEl.textContent = `Update check failed (${why}) — you may not be on the latest version.`;
+    hideUpdateBanner();
+  } else if (checked) {
+    aboutStatusEl.textContent = stamp
+      ? `You're up to date. Last checked ${stamp}.`
+      : "You're up to date.";
+    hideUpdateBanner();
+  } else {
+    aboutStatusEl.textContent = 'Not checked yet.';
+  }
+}
+
+function showUpdateBanner(info) {
+  if (!updateBanner) return;
+  updateBanner.style.display = '';
+  if (updateLatestEl) updateLatestEl.textContent = `v${info.latestVersion}`;
+  if (updateReleasedEl) {
+    const released = info.publishedAt ? formatRelative(new Date(info.publishedAt)) : '';
+    updateReleasedEl.textContent = released ? `— Released ${released}` : '';
+  }
+  if (updateLinkEl) updateLinkEl.setAttribute('href', info.releaseUrl || '#');
+}
+
+function hideUpdateBanner() {
+  if (updateBanner) updateBanner.style.display = 'none';
+}
+
+function hideSidebarDot() {
+  const dot = document.getElementById('update-dot');
+  if (dot) dot.style.display = 'none';
+  const banner = document.getElementById('sidebar-update-banner');
+  if (banner) banner.style.display = 'none';
+}
+
+function formatRelative(date) {
+  const diffMs = Date.now() - date.getTime();
+  const min = Math.floor(diffMs / 60000);
+  if (min < 1) return 'just now';
+  if (min < 60) return `${min} min ago`;
+  const hours = Math.floor(min / 60);
+  if (hours < 24) return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  return `${days}d ago`;
+}
+
+async function syncToggleFromSettings() {
+  if (!toggleEl) return;
+  const value = await ipcRenderer.invoke(IPC.GET_USER_SETTING, TELEMETRY_KEY);
+  // Default ON when unset (opt-out semantics)
+  toggleEl.checked = value !== false;
+
+  if (crashDumpsToggleEl) {
+    const dumps = await ipcRenderer.invoke(IPC.GET_USER_SETTING, CRASH_DUMPS_KEY);
+    // Default ON when unset (local-only; nothing is uploaded)
+    crashDumpsToggleEl.checked = dumps !== false;
+  }
+}
+
+module.exports = {
+  init,
+  open: () => overlay && overlay.open(),
+  close: () => overlay && overlay.close(),
+  toggle: () => overlay && overlay.toggle()
+};
