--- src/renderer/settingsModal.js
@@ -1,506 +1,0 @@
-/**
- * Settings Modal
- *
- * Minimal preferences UI. v1 ships a single "Privacy & Analytics" section
- * with the telemetry toggle. Future settings (theme, AI tool defaults,
- * keybinding remap) plug into this same modal.
- */
-
-const { ipcRenderer, shell } = require('electron');
-const { IPC } = require('../shared/ipcChannels');
-const state = require('./state');
-const specDrivenHint = require('./specDrivenHint');
-
-const TELEMETRY_KEY = 'telemetryEnabled';
-const CRASH_DUMPS_KEY = 'crashDumpsEnabled';
-const DISMISSED_VERSION_KEY = 'dismissedUpdateVersion';
-
-let overlayEl = null;
-let toggleEl = null;
-let crashDumpsToggleEl = null;
-let specDrivenToggleEl = null;
-let gitSharingSelectEl = null;
-let gitSharingWarningEl = null;
-let removeFrameBtnEl = null;
-let removeFrameNoteEl = null;
-let specDrivenNoteEl = null;
-let isOpen = false;
-
-// About section elements + state
-let aboutVersionEl = null;
-let aboutStatusEl = null;
-let aboutCheckBtn = null;
-let updateBanner = null;
-let updateLatestEl = null;
-let updateReleasedEl = null;
-let updateLinkEl = null;
-let updateDismissBtn = null;
-let currentUpdateInfo = null;
-
-function init() {
-  overlayEl = document.getElementById('settings-overlay');
-  toggleEl = document.getElementById('settings-telemetry-toggle');
-  crashDumpsToggleEl = document.getElementById('settings-crash-dumps-toggle');
-  specDrivenToggleEl = document.getElementById('settings-spec-driven-toggle');
-  specDrivenNoteEl = document.getElementById('settings-spec-driven-note');
-  gitSharingSelectEl = document.getElementById('settings-git-sharing');
-  gitSharingWarningEl = document.getElementById('settings-git-sharing-warning');
-  removeFrameBtnEl = document.getElementById('settings-remove-frame');
-  removeFrameNoteEl = document.getElementById('settings-remove-frame-note');
-
-  // About section
-  aboutVersionEl = document.getElementById('settings-version');
-  aboutStatusEl = document.getElementById('settings-update-status');
-  aboutCheckBtn = document.getElementById('settings-check-updates');
-  updateBanner = document.getElementById('settings-update-available');
-  updateLatestEl = document.getElementById('settings-update-latest');
-  updateReleasedEl = document.getElementById('settings-update-released');
-  updateLinkEl = document.getElementById('settings-update-link');
-  updateDismissBtn = document.getElementById('settings-update-dismiss');
-
-  if (!overlayEl || !toggleEl) {
-    console.error('Settings modal: required elements not found');
-    return;
-  }
-
-  // Load current value
-  syncToggleFromSettings();
-  initAboutSection();
-
-  // Toggle: persist + tell main process to enable/disable Aptabase
-  toggleEl.addEventListener('change', async () => {
-    const enabled = toggleEl.checked;
-    await ipcRenderer.invoke(IPC.SET_USER_SETTING, TELEMETRY_KEY, enabled);
-    await ipcRenderer.invoke(IPC.TELEMETRY_SET_ENABLED, enabled);
-  });
-
-  // Crash dumps: persisted setting only — crashGuard reads it at startup
-  // (the reporter can't be stopped once started, so changes apply on next launch)
-  if (crashDumpsToggleEl) {
-    crashDumpsToggleEl.addEventListener('change', async () => {
-      await ipcRenderer.invoke(IPC.SET_USER_SETTING, CRASH_DUMPS_KEY, crashDumpsToggleEl.checked);
-    });
-  }
-
-  // Spec-Driven Development: per-project flag in .frame/config.json, not a
-  // user preference — main writes the config and AGENTS.md section. On
-  // failure we snap the switch back so it never lies about the on-disk state.
-  if (specDrivenToggleEl) {
-    specDrivenToggleEl.addEventListener('change', async () => {
-      const projectPath = state.getProjectPath();
-      const wanted = specDrivenToggleEl.checked;
-      if (!projectPath) {
-        specDrivenToggleEl.checked = !wanted;
-        return;
-      }
-      specDrivenToggleEl.disabled = true;
-      try {
-        const result = await ipcRenderer.invoke(IPC.SET_SPEC_DRIVEN, {
-          projectPath,
-          enabled: wanted
-        });
-        if (!result || !result.success) {
-          specDrivenToggleEl.checked = !wanted;
-          setSpecDrivenNote(
-            'Could not change this setting: ' + ((result && result.error) || 'unknown error')
-          );
-        } else {
-          setSpecDrivenNote(null);
-          // Turning it off here is a deliberate choice — stop offering to
-          // turn it back on for this project.
-          if (!wanted) await specDrivenHint.markDismissed(projectPath);
-          specDrivenHint.refresh();
-        }
-      } catch (err) {
-        specDrivenToggleEl.checked = !wanted;
-        setSpecDrivenNote('Could not change this setting: ' + err.message);
-      } finally {
-        specDrivenToggleEl.disabled = false;
-      }
-    });
-  }
-
-  // Git sharing: per-project, like the spec-driven flag. Main applies the
-  // mode (exclude block, .frame/.gitignore, which settings file the hooks
-  // live in) and answers with the state we re-render from — including the
-  // warning when .frame/ is already committed.
-  if (gitSharingSelectEl) {
-    gitSharingSelectEl.addEventListener('change', async () => {
-      const projectPath = state.getProjectPath();
-      const mode = gitSharingSelectEl.value;
-      if (!projectPath) return;
-      gitSharingSelectEl.disabled = true;
-      try {
-        const result = await ipcRenderer.invoke(IPC.SET_GIT_SHARING, { projectPath, mode });
-        renderGitSharing(result);
-      } catch (err) {
-        setGitSharingWarning('Could not change this setting: ' + err.message);
-      } finally {
-        gitSharingSelectEl.disabled = false;
-      }
-    });
-  }
-
-  // Remove Frame: destructive and irreversible, so it confirms first and
-  // says exactly what will be deleted. Main does the work and answers with a
-  // list; the note reports it rather than a bare "done".
-  if (removeFrameBtnEl) {
-    removeFrameBtnEl.addEventListener('click', async () => {
-      const projectPath = state.getProjectPath();
-      if (!projectPath) {
-        setRemoveFrameNote('Open a project first.');
-        return;
-      }
-      const confirmed = window.confirm(
-        'Remove Frame from this project?\n\n' +
-        'This deletes .frame/ (including your specs, notes and tasks) and ' +
-        '.claude/rules/frame.md, takes Frame\'s hook entries out of ' +
-        '.claude/settings.json and settings.local.json — deleting either file ' +
-        'if Frame\'s entries were all it held — takes Frame\'s block out of the ' +
-        'pre-commit hook, and removes its block from .git/info/exclude. ' +
-        'Anything you wrote is left where it is.\n\n' +
-        'This cannot be undone.'
-      );
-      if (!confirmed) return;
-
-      removeFrameBtnEl.disabled = true;
-      try {
-        const result = await ipcRenderer.invoke(IPC.REMOVE_FRAME_FROM_PROJECT, projectPath);
-        if (result && result.errors && result.errors.length > 0) {
-          setRemoveFrameNote('Removed, with problems: ' + result.errors.join('; '));
-        } else {
-          setRemoveFrameNote('Removed: ' + ((result && result.removed) || []).join(', '));
-        }
-        // The project is no longer a Frame project: say so, or the spec
-        // panel keeps offering to write into a .frame/ that is gone.
-        state.noteFrameRemoved(projectPath);
-        await syncGitSharing();
-        await syncSpecDrivenToggle();
-      } catch (err) {
-        setRemoveFrameNote('Could not remove Frame: ' + err.message);
-      } finally {
-        removeFrameBtnEl.disabled = false;
-      }
-    });
-  }
-
-  overlayEl.addEventListener('mousedown', (e) => {
-    if (e.target === overlayEl) close();
-  });
-
-  document.querySelectorAll('[data-settings-close]').forEach((btn) => {
-    btn.addEventListener('click', () => close());
-  });
-
-  document.addEventListener('keydown', (e) => {
-    if (isOpen && e.key === 'Escape') {
-      e.preventDefault();
-      close();
-    }
-  });
-
-  // Open from menu trigger
-  ipcRenderer.on(IPC.OPEN_SETTINGS, () => open());
-
-  // Push updates from periodic recheck refresh the About panel state
-  ipcRenderer.on(IPC.UPDATE_AVAILABLE, (event, info) => {
-    currentUpdateInfo = info;
-    renderUpdateState({ checked: true, found: true, info });
-  });
-}
-
-function initAboutSection() {
-  // Version text from package.json
-  try {
-    const pkgVersion = require('../../package.json').version;
-    if (aboutVersionEl) aboutVersionEl.textContent = `v${pkgVersion}`;
-  } catch (e) { /* ignore */ }
-
-  if (aboutCheckBtn) {
-    aboutCheckBtn.addEventListener('click', () => {
-      runCheck(true);
-    });
-  }
-
-  // Diagnostics: reveal the rotating (redacted) log file — the first thing
-  // to grab when filing a bug report. Location documented in PRIVACY.md.
-  const openLogsBtn = document.getElementById('settings-open-logs');
-  if (openLogsBtn) {
-    openLogsBtn.addEventListener('click', async () => {
-      try {
-        const info = await ipcRenderer.invoke(IPC.GET_LOG_INFO);
-        if (info && info.logPath) shell.showItemInFolder(info.logPath);
-        else if (info && info.logsDir) shell.openPath(info.logsDir);
-      } catch (err) {
-        console.error('Settings: could not open logs folder', err);
-      }
-    });
-  }
-
-  if (updateLinkEl) {
-    updateLinkEl.addEventListener('click', (e) => {
-      e.preventDefault();
-      if (currentUpdateInfo && currentUpdateInfo.releaseUrl) {
-        shell.openExternal(currentUpdateInfo.releaseUrl);
-      }
-    });
-  }
-
-  if (updateDismissBtn) {
-    updateDismissBtn.addEventListener('click', async () => {
-      if (!currentUpdateInfo) return;
-      await ipcRenderer.invoke(
-        IPC.SET_USER_SETTING,
-        DISMISSED_VERSION_KEY,
-        currentUpdateInfo.latestVersion
-      );
-      // Hide the banner immediately; sidebar dot is also gated by this flag.
-      hideUpdateBanner();
-      hideSidebarDot();
-    });
-  }
-
-  // Hydrate from main's cached status (no extra network call)
-  ipcRenderer
-    .invoke(IPC.GET_UPDATE_STATUS)
-    .then((status) => {
-      if (!status) return;
-      if (status.result) {
-        currentUpdateInfo = status.result;
-        renderUpdateState({
-          checked: !!status.lastCheckedAt,
-          found: true,
-          info: status.result,
-          checkedAt: status.lastCheckedAt
-        });
-      } else if (status.lastStatus === 'error') {
-        renderUpdateState({
-          checked: true,
-          failed: true,
-          failReason: status.lastErrorReason,
-          checkedAt: status.lastCheckedAt
-        });
-      } else if (status.lastCheckedAt) {
-        renderUpdateState({
-          checked: true,
-          found: false,
-          checkedAt: status.lastCheckedAt
-        });
-      } else {
-        // Not yet checked since launch — fire one to populate
-        runCheck(false);
-      }
-    })
-    .catch(() => {});
-}
-
-async function runCheck(userInitiated) {
-  if (aboutStatusEl) aboutStatusEl.textContent = 'Checking…';
-  if (aboutCheckBtn) aboutCheckBtn.disabled = true;
-  try {
-    // Discriminated result: 'update-available' | 'up-to-date' | 'error'.
-    // A failed check must never render as "you're up to date".
-    const res = await ipcRenderer.invoke(IPC.CHECK_FOR_UPDATE);
-    if (res && res.status === 'update-available') {
-      currentUpdateInfo = res.info;
-      renderUpdateState({
-        checked: true,
-        found: true,
-        info: res.info,
-        checkedAt: res.checkedAt,
-        userInitiated
-      });
-    } else if (res && res.status === 'error') {
-      renderUpdateState({
-        checked: true,
-        failed: true,
-        failReason: res.reason,
-        checkedAt: res.checkedAt,
-        userInitiated
-      });
-    } else {
-      renderUpdateState({
-        checked: true,
-        found: false,
-        checkedAt: res ? res.checkedAt : null,
-        userInitiated
-      });
-    }
-  } catch (err) {
-    if (aboutStatusEl) aboutStatusEl.textContent = 'Could not check for updates.';
-  } finally {
-    if (aboutCheckBtn) aboutCheckBtn.disabled = false;
-  }
-}
-
-function renderUpdateState({ checked, found, failed, failReason, info, checkedAt, userInitiated }) {
-  if (!aboutStatusEl) return;
-  const stamp = checkedAt ? formatRelative(new Date(checkedAt)) : '';
-  if (found && info) {
-    aboutStatusEl.textContent = stamp ? `Last checked ${stamp}.` : '';
-    showUpdateBanner(info);
-  } else if (failed) {
-    const why = failReason === 'timeout'
-      ? 'timed out'
-      : failReason === 'parse'
-        ? 'unexpected response'
-        : 'network error';
-    aboutStatusEl.textContent = `Update check failed (${why}) — you may not be on the latest version.`;
-    hideUpdateBanner();
-  } else if (checked) {
-    aboutStatusEl.textContent = stamp
-      ? `You're up to date. Last checked ${stamp}.`
-      : "You're up to date.";
-    hideUpdateBanner();
-  } else {
-    aboutStatusEl.textContent = 'Not checked yet.';
-  }
-}
-
-function showUpdateBanner(info) {
-  if (!updateBanner) return;
-  updateBanner.style.display = '';
-  if (updateLatestEl) updateLatestEl.textContent = `v${info.latestVersion}`;
-  if (updateReleasedEl) {
-    const released = info.publishedAt ? formatRelative(new Date(info.publishedAt)) : '';
-    updateReleasedEl.textContent = released ? `— Released ${released}` : '';
-  }
-  if (updateLinkEl) updateLinkEl.setAttribute('href', info.releaseUrl || '#');
-}
-
-function hideUpdateBanner() {
-  if (updateBanner) updateBanner.style.display = 'none';
-}
-
-function hideSidebarDot() {
-  const dot = document.getElementById('update-dot');
-  if (dot) dot.style.display = 'none';
-  const banner = document.getElementById('sidebar-update-banner');
-  if (banner) banner.style.display = 'none';
-}
-
-function formatRelative(date) {
-  const diffMs = Date.now() - date.getTime();
-  const min = Math.floor(diffMs / 60000);
-  if (min < 1) return 'just now';
-  if (min < 60) return `${min} min ago`;
-  const hours = Math.floor(min / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
-}
-
-async function syncToggleFromSettings() {
-  if (!toggleEl) return;
-  const value = await ipcRenderer.invoke(IPC.GET_USER_SETTING, TELEMETRY_KEY);
-  // Default ON when unset (opt-out semantics)
-  toggleEl.checked = value !== false;
-
-  if (crashDumpsToggleEl) {
-    const dumps = await ipcRenderer.invoke(IPC.GET_USER_SETTING, CRASH_DUMPS_KEY);
-    // Default ON when unset (local-only; nothing is uploaded)
-    crashDumpsToggleEl.checked = dumps !== false;
-  }
-
-  await syncSpecDrivenToggle();
-  await syncGitSharing();
-}
-
-/**
- * Reflect the current project's features.specDriven flag. With no project
- * open (or a folder that isn't a Frame project yet) there is nothing to
- * write, so the switch is disabled and says why.
- */
-async function syncSpecDrivenToggle() {
-  if (!specDrivenToggleEl) return;
-  const projectPath = state.getProjectPath();
-  if (!projectPath) {
-    specDrivenToggleEl.checked = false;
-    specDrivenToggleEl.disabled = true;
-    setSpecDrivenNote('Open a project to change this — the setting lives in its .frame/config.json.');
-    return;
-  }
-  try {
-    const enabled = await ipcRenderer.invoke(IPC.IS_SPEC_DRIVEN_ENABLED, projectPath);
-    specDrivenToggleEl.checked = enabled === true;
-    specDrivenToggleEl.disabled = false;
-    setSpecDrivenNote(null);
-  } catch (err) {
-    specDrivenToggleEl.disabled = true;
-    setSpecDrivenNote('Could not read this project’s Frame config.');
-  }
-}
-
-/**
- * Paint the sharing row from main's state object. `error` means the project
- * isn't a Frame project (or none is open) — the control goes inert rather
- * than showing a mode the project doesn't have.
- */
-function renderGitSharing(stateObj) {
-  if (!gitSharingSelectEl) return;
-  if (!stateObj || stateObj.error) {
-    gitSharingSelectEl.disabled = true;
-    setGitSharingWarning(
-      stateObj && stateObj.error === 'not a Frame project'
-        ? 'Initialize this project with Frame to choose how its files relate to git.'
-        : null
-    );
-    return;
-  }
-  gitSharingSelectEl.disabled = false;
-  gitSharingSelectEl.value = stateObj.mode;
-  setGitSharingWarning(stateObj.warning);
-}
-
-async function syncGitSharing() {
-  if (!gitSharingSelectEl) return;
-  const projectPath = state.getProjectPath();
-  if (!projectPath) {
-    renderGitSharing({ error: 'no project' });
-    return;
-  }
-  try {
-    renderGitSharing(await ipcRenderer.invoke(IPC.GET_GIT_SHARING_STATE, projectPath));
-  } catch (err) {
-    renderGitSharing({ error: err.message });
-  }
-}
-
-function setRemoveFrameNote(message) {
-  if (!removeFrameNoteEl) return;
-  removeFrameNoteEl.textContent = message || '';
-  removeFrameNoteEl.style.display = message ? '' : 'none';
-}
-
-function setGitSharingWarning(message) {
-  if (!gitSharingWarningEl) return;
-  gitSharingWarningEl.textContent = message || '';
-  gitSharingWarningEl.style.display = message ? '' : 'none';
-}
-
-function setSpecDrivenNote(message) {
-  if (!specDrivenNoteEl) return;
-  specDrivenNoteEl.textContent = message || '';
-  specDrivenNoteEl.style.display = message ? '' : 'none';
-}
-
-function open() {
-  if (isOpen) return;
-  isOpen = true;
-  syncToggleFromSettings();
-  overlayEl.classList.add('visible');
-}
-
-function close() {
-  if (!isOpen) return;
-  isOpen = false;
-  overlayEl.classList.remove('visible');
-  if (typeof window.terminalFocus === 'function') window.terminalFocus();
-}
-
-function toggle() {
-  if (isOpen) close();
-  else open();
-}
-
-module.exports = { init, open, close, toggle };
